♻️ refactor : 전역 상태에서 user_id 제거
